+++ src/tilemap.go
@@ -90,10 +90,10 @@ func (o *ObjectGroup) GetObjectByName(name string) (*Object, error){
 
 
 type Map struct {
-	Width        uint16        `xml:"width,attr"`      // Number of tiles Width
-	Height       uint16       `xml:"height,attr"`     // Number of tiles Height
-	TileWidth    uint16       `xml:"tilewidth,attr"`  // Tile Width size in pixels
-	TileHeight   uint16       `xml:"tileheight,attr"` // Tile Height size in pixels
+	Width        int32        `xml:"width,attr"`      // Number of tiles Width
+	Height       int32       `xml:"height,attr"`     // Number of tiles Height
+	TileWidth    int32       `xml:"tilewidth,attr"`  // Tile Width size in pixels
+	TileHeight   int32       `xml:"tileheight,attr"` // Tile Height size in pixels
 	Infinite     bool         `xml:"infinite,attr"`
 	RenderOrder  string       `xml:"renderorder,attr"`
 	Layers       []*Layer     `xml:"layer"`
